revert - float32 - seems more trouble than I can confirm its worth.  we profiled the memory usage of ward_tree() (a major contributor to memory use), which must cast to float 64 under the hood somewhere back from 32 ... (we increased num voxels on a WGN cube and noticed that the slope of the line (increase in MB / vox) was steeper on the float32 input than the float64).

+++ glow/experiment/exper.py
@@ -60,7 +60,7 @@ class ExperimentImageOnly:
 
         num_vox = np.prod(shape)
         rng = np.random.default_rng(seed=seed)
-        y = rng.multivariate_normal(np.zeros(b), np.eye(b), num_img * num_vox).astype(np.float32)
+        y = rng.multivariate_normal(np.zeros(b), np.eye(b), num_img * num_vox)
 
         # reshape, de-mean, reshape
         y = y.reshape((b, num_img, num_vox))
@@ -139,7 +139,7 @@ class ExperimentImageOnly:
         # mask into each image, store as y
         mask = mask_idx >= 0
         y_names = sorted(feat_sbj_img.keys())
-        y = np.empty((len(feat_sbj_img), df.shape[0], mask.sum()), dtype=np.float32)
+        y = np.empty((len(feat_sbj_img), df.shape[0], mask.sum()))
         for sbj_idx, sbj in enumerate(sorted(df.index)):
             for feat_idx, feat in enumerate(y_names):
                 y[feat_idx, sbj_idx, :] = feat_sbj_img[feat][sbj][mask]
